Relocate general query and mutation tests to graphql crate
