+++ Cargo.toml
@@ -16,7 +16,6 @@ members = [
     "graphql/invoice",
     "graphql/invoice_line",
     "graphql/location",
-    "graphql/general_queries",
-    "graphql/general_mutations",
+    "graphql/general",
     "graphql/batch_mutations",
 ]
