+++ graphql/Cargo.toml
@@ -19,8 +19,7 @@ service = { path = "../service" }
 util = { path = "../util" }
 graphql_core = { path = "core" }
 graphql_types = { path = "types" }
-graphql_general_queries = { path = "general_queries" }
-graphql_general_mutations = { path = "general_mutations" }
+graphql_general = { path = "general" }
 graphql_location = { path = "location" }
 graphql_invoice = { path = "invoice" }
 graphql_invoice_line = { path = "invoice_line" }
@@ -29,7 +28,6 @@ graphql_requisition_line = { path = "requisition_line" }
 graphql_stocktake = { path = "stocktake" }
 graphql_stocktake_line = { path = "stocktake_line" }
 graphql_batch_mutations = { path = "batch_mutations" }
-
 
 actix-web = "3.3.2" # Versions >=v4 depend on Tokio v1.
 anymap = "0.12"
@@ -41,6 +39,11 @@ reqwest = { version = "0.10", features = ["json"] } # Versions >=0.11 depend on 
 serde = "1.0.126"
 serde_json = "1.0.66"
 thiserror = "1.0.30"
+
+
+[dev-dependencies]
+actix-rt = "1.1.1" # for Tokio 0.2
+assert-json-diff = "2.0.1"
 
 [features]
 default = ["repository/sqlite"]
