+++ graphql/general/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "graphql_general_queries"
+name = "graphql_general"
 version = "0.1.0"
 edition = "2018"
 
@@ -24,6 +24,12 @@ reqwest = { version = "0.10", features = ["json"] } # Versions >=0.11 depend on 
 serde = "1.0.126"
 serde_json = "1.0.66"
 thiserror = "1.0.30"
+
+
+[dev-dependencies]
+actix-rt = "1.1.1" # for Tokio 0.2
+assert-json-diff = "2.0.1"
+
 
 [features]
 default = ["repository/sqlite"]
