+++ graphql/general/src/mutations/mod.rs
@@ -1,0 +1,1 @@
+pub mod user_register;
