+++ graphql/general/src/queries/invoice_counts.rs
