+++ graphql/general/src/queries/logout.rs
