+++ graphql/general/src/queries/master_list.rs
