+++ graphql/general/src/queries/mod.rs
@@ -17,3 +17,7 @@ pub use self::item::*;
 pub mod stock_counts;
 pub use self::stock_counts::*;
 pub mod store;
+pub use self::store::*;
+
+#[cfg(test)]
+mod tests;
