+++ graphql/general/src/queries/stock_counts.rs
