+++ graphql/general/src/queries/store.rs
