+++ graphql/general/src/queries/tests/mod.rs
@@ -1,0 +1,4 @@
+mod item_stats;
+mod items;
+mod master_lists;
+mod names;
