--- graphql/general_mutations/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "graphql_general_mutations"
-version = "0.1.0"
-edition = "2018"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-
-repository = { path = "../../repository" }
-service = { path = "../../service" }
-util = { path = "../../util" }
-graphql_core = { path = "../core" }
-graphql_types = { path = "../types" }
-
-actix-web = "3.3.2" # Versions >=v4 depend on Tokio v1.
-anymap = "0.12"
-async-graphql = { version = "=2.11.3", features = ["dataloader", "chrono"] }
-async-graphql-actix-web = "=2.11.3"
-async-trait = "0.1.16"
-chrono = { version = "0.4", features = ["serde"] }
-reqwest = { version = "0.10", features = ["json"] } # Versions >=0.11 depend on Tokio v1.
-serde = "1.0.126"
-serde_json = "1.0.66"
-thiserror = "1.0.30"
-
-[features]
-default = ["repository/sqlite"]
-postgres = ["repository/postgres"]
