--- graphql/general_mutations/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod user_register;
-pub use self::user_register::*;
